+++ R/helpers.R
@@ -108,3 +108,39 @@ sg_kill <- function(sg){
   sg$x$kill <- TRUE
   sg
 }
+
+#' Color
+#' 
+#' Scale color by node size.
+#' 
+#' @inheritParams sg_nodes
+#' @param pal Vector of color.
+#' 
+#' @examples 
+#' nodes <- sg_make_nodes() 
+#' edges <- sg_make_edges(nodes, 20)
+#' 
+#' sigmajs() %>% 
+#'   sg_nodes(nodes, id, size) %>% 
+#'   sg_scale_color(pal = c("red", "blue"))
+#' 
+#' @name color-scale
+#' @export
+sg_scale_color <- function(sg, pal){
+  
+  size <- purrr::map(sg$x$data$nodes, "size") %>% 
+    unlist() %>% 
+    as.numeric()
+  
+  if(!length(size))
+    stop("no node size passed", call. = FALSE)
+  
+  color <- scales::col_numeric(pal, domain = range(size))(size)
+  
+  sg$x$data$nodes <- purrr::map2(sg$x$data$nodes, color, function(x, y){
+    x$color <- y
+    return(x)
+  })
+  
+  return(sg)
+}